// src/api/schemas/discussionSummaryOut.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * MyApp API
 * OpenAPI spec version: 1.0.0
 */
import type { DiscussionSummaryOutCreatedBy } from './discussionSummaryOutCreatedBy';
import type { DiscussionSummaryOutId } from './discussionSummaryOutId';
import type { DiscussionSummaryOutLastUpdatedBy } from './discussionSummaryOutLastUpdatedBy';

export interface DiscussionSummaryOut {
  /** Markdown/text summary of the discussions */
  content: string;
  created_at: string;
  created_by?: DiscussionSummaryOutCreatedBy;
  id?: DiscussionSummaryOutId;
  last_updated_by?: DiscussionSummaryOutLastUpdatedBy;
  updated_at: string;
}